+++ src/operations.cc
@@ -27,29 +27,6 @@ using vips::VImage;
 using vips::VError;
 
 namespace sharp {
-
-  /*
-    Removes alpha channel, if any.
-  */
-  VImage RemoveAlpha(VImage image) {
-    if (HasAlpha(image)) {
-      image = image.extract_band(0, VImage::option()->set("n", image.bands() - 1));
-    }
-    return image;
-  }
-
-  /*
-    Ensures alpha channel, if missing.
-  */
-  VImage EnsureAlpha(VImage image) {
-    if (!HasAlpha(image)) {
-      std::vector<double> alpha;
-      alpha.push_back(sharp::MaximumImageAlpha(image.interpretation()));
-      image = image.bandjoin_const(alpha);
-    }
-    return image;
-  }
-
   /*
    * Tint an image using the specified chroma, preserving the original image luminance
    */